Added Mv member reflect_in_blade.  Fixed printing for Dop's and Sdop's. Fixed latex printing for basis bases and blades (now they print boldsymbol). Updated docs to reflect changes and corrected some mistakes in docs.

+++ galgebra/setgapth.py
@@ -1,6 +1,13 @@
 #!/usr/bin/python
 #setgapth.py
 
+Completion_Message = \
+"""
+Python pth file Ga.pth for all modules of galgebra has been created and
+placed in directory appropriate for Linux, Windows, or OSX.  Note that
+if Ga.pth already exists the old version will be deleted.
+"""
+
 import os, sys, site
 
 cur_dir = os.path.abspath('.')
@@ -8,14 +15,15 @@ cur_dir = os.path.abspath('.')
 if os.name == 'posix' or os.name == 'mac':
     dist_pkgs = site.getsitepackages()[-1]
     pth_name = dist_pkgs + '/Ga.pth'
-    os.system('rm -f ' + pth_name)
+    if os.path.exists(pth_name):
+        os.system('rm -f ' + pth_name)
 
 if os.name == 'nt':
     dist_pkgs = site.getsitepackages()[-1]
     dist_lst = dist_pkgs.split('\\')
     pth_name = dist_lst[0] + '\\' + dist_lst[1] + '\\Ga.pth'
-    #pth_name = 'C:\\Python27\\Ga.pth'
-    os.system('del ' + pth_name)
+    if os.path.exists(pth_name):
+        os.system('del ' + pth_name)
 
 pth_file = open(pth_name,'w')
 pth_file.write('#Path of Ga module\n'+cur_dir)
@@ -25,5 +33,7 @@ print 'os name:',os.name
 print 'site-packages directory:',pth_name
 print 'Ga.pth:'
 print os.system('more ' + pth_name)
+
+print Completion_Message
